refactor: fix 20 clippy warnings, port Yazi scrollable algorithm
- Replace homegrown scroll with yazi-widgets Scrollable::next/prev port
- scrolloff = (visible_height / 2).min(5) for small/large windows
- Derive Default for SettingsSnapshot and ControlChannel
- Replace manual Option::map with combinator chain in main.rs
- Remove redundant closures, double &PathBuf params, useless format!
- Fix to_vec(), div_ceil(), identical if blocks, unused variable

--------------------------------------------------

refactor: 修复 20 个 clippy 警告，移植 Yazi Scrollable 算法

- 用 yazi-widgets 的 next/prev 算法替换自研滚动
- scrolloff = (visible_height / 2).min(5)
- SettingsSnapshot 和 ControlChannel 派生 Default
- main.rs 手动 Option::map 改为组合链

+++ crates/horikawa-tui/src/views/visualizer.rs
@@ -103,7 +103,7 @@ pub fn draw_vis_bars(lines: &mut Vec<Line<'static>>, data: &[f32], height: usize
         1
     };
     let gap = 1u16;
-    let total_per_bar = (bw + gap) as u16;
+    let total_per_bar = bw + gap;
     let num_bars = (width as u16 / total_per_bar)
         .min(data.len().max(1) as u16)
         .max(1) as usize;
@@ -162,7 +162,7 @@ pub fn draw_vis_spectrum(lines: &mut Vec<Line<'static>>, data: &[f32], height: u
         1
     };
     let gap = 1u16;
-    let total_per_bar = (bw + gap) as u16;
+    let total_per_bar = bw + gap;
     let num_bars = (width as u16 / total_per_bar)
         .min(data.len().max(1) as u16)
         .max(1) as usize;
